:recycle: refactor: 컨텐츠 신쳥 완료 페이지 컴포넌트 분리
